Extracted the guts of the scripts into a library.
This entailed restructuring the per-Estate collections, a bit, to make them easier to reason about and to make the script bodies a bit more regular. The old structure:

* `collection`: the root of an event collection
    * `collection/TIMESTAMP`: document link
    * `collection/TIMESTAMP.yaml`: event metadata

Hard to parse with `readdir` and friends, because each entry is paired.

New structure:

* `collection`: still the root of an event collection
    * `collection/TIMESTAMP`: the root of a single event
        * `collection/TIMESTAMP/document`: document link
        * `collection/TIMESTAMP/event.yaml`: event metadata

No more pairs.

The existing collections have been rearranged accordingly.